fix(ex00): change type

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: agautier <agautier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/29 12:10:59 by agautier          #+#    #+#             */
-/*   Updated: 2021/10/29 15:37:13 by agautier         ###   ########.fr       */
+/*   Updated: 2021/10/30 17:45:02 by agautier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	main(int argc, char **argv)
 		std::cout << "* LOURD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
 		return (EXIT_SUCCESS);
 	}
-	for (int i = 1; i < argc; i++)
+	for (uint8_t i = 1; i < argc; i++)
 		for (size_t j = 0; j < strlen(argv[i]); j++)
 			std::cout << (char)std::toupper(argv[i][j]);
 	std::cout << std::endl;
